functions/planning/requirements.txt
attrs==19.3.0
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
clickclick==1.2.2
Flask==1.1.2
geographiclib==1.50
geopy==1.22.0
google-api-core==1.17.0
google-auth==1.14.3
google-cloud-core==1.3.0
google-cloud-datastore==1.12.0
googleapis-common-protos==1.51.0
googlemaps==4.4.0
grpcio==1.29.0
idna==2.9
importlib-metadata==1.6.0
inflection==0.4.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
openapi-spec-validator==0.2.8
ortools==7.6.7691
protobuf==3.12.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyrsistent==0.16.0
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
rsa==4.0
six==1.14.0
urllib3==1.25.9
Werkzeug==1.0.1
zipp==3.1.0
